requirements.txt
certifi==2021.10.8
cycler==0.11.0
dominate==2.6.0
fonttools==4.29.1
imageio==2.16.0
kiwisolver==1.3.2
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scipy==1.8.0
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
torch==1.10.2
torchsummary==1.5.1
torchvision==0.11.3
tqdm==4.62.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
